server

+++ protocol/include/protocol/login/Login.h
@@ -5,7 +5,7 @@
 typedef struct
 {
     char password[16];
-    uint16_t username_len;
+    uint8_t username_len;
     char username[0];
 } CRPPacketLogin;
